requirements.txt
torch
torchaudio
numpy
pandas
tqdm
pyyaml
